handling new symfony strict type checks in FixturesKernel
